requirements.txt
aiocache==0.11.1
aiohttp==3.7.4.post0
alembic==1.6.5
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.6
redis
hiredis
chardet==3.0.4
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
coloredlogs==15.0.1
configparser==5.0.2
cycler==0.10.0
dacite==1.6.0
dbmanager @ git+https://github.com/yupix/dbmanager@93258484bcfe90a6ca10850cb121cfb993ec1039
git+https://github.com/yupix/alfakana.git
python-i18n[YAML]
discord-py-slash-command==2.3.2
discord.py==1.7.3
fastapi==0.67.0
fastapi-discord==0.1.6
fastapi-versioning==0.9.1
feedparser==6.0.8
ffmpeg==1.4
flake8==3.9.2
googletrans==4.0.0rc1
greenlet==1.1.0
h11==0.9.0
h2==3.2.0
halo==0.0.31
hpack==3.0.0
hstspreload==2021.7.5
httpcore==0.9.1
httpx==0.13.3
humanfriendly==9.2
hyperframe==5.2.0
idna==2.10
inputimeout==1.0.4
kiwisolver==1.3.1
log-symbols==0.0.14
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==5.1.0
numpy==1.21.1
Pillow==8.3.1
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyflakes==2.3.1
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
pytimedinput==1.3.1
PyYAML==5.4.1
requests==2.26.0
rfc3986==1.5.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
spinners==0.0.24
SQLAlchemy==1.4.22
starlette==0.14.2
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.6
uvicorn==0.14.0
yarl==1.6.3
